chore(suite-common): Resolve circular dependencies in suite-common/trading types
