+++ suite-common/trading/src/reducers/tradingReducer.ts
@@ -75,6 +75,12 @@ export interface TradingState {
     settings: TradingSettingsState;
 }
 
+export type TradingRootState = {
+    wallet: {
+        trading: TradingState;
+    };
+};
+
 export const initialState: TradingState = {
     info: {
         platforms: undefined,
