+++ suite-common/trading/src/selectors/__tests__/tradingSelectors.test.ts
@@ -17,14 +17,10 @@ import { accountBtc, accountEth } from '../../__fixtures__/utils';
 import { BuyInfo, TradingBuyState } from '../../reducers/buyReducer';
 import { ExchangeInfo, exchangeInitialState } from '../../reducers/exchangeReducer';
 import { SellInfo, sellInitialState } from '../../reducers/sellReducer';
-import { initialState } from '../../reducers/tradingReducer';
+import { type TradingRootState, initialState } from '../../reducers/tradingReducer';
+import type { TradingPaymentMethodListProps, TradingPaymentMethodProps } from '../../types';
 import {
-    TradingPaymentMethodListProps,
-    TradingPaymentMethodProps,
-    TradingRootState,
     TradingRootStateWithDeviceAndAccounts,
-} from '../../types';
-import {
     selectBestBuyQuoteByPaymentMethod,
     selectBuyQuotesByPaymentMethod,
     selectDeviceHasTradingTrades,
