+++ suite-common/trading/src/selectors/settingsSelectors.ts
@@ -1,5 +1,5 @@
 import { TRADING_SETTINGS_MAX_SLIPPAGE_PERCENTAGE_DEFAULT } from '../constants';
-import { TradingRootState } from '../types';
+import type { TradingRootState } from '../reducers/tradingReducer';
 
 export const selectTradingMaxSlippagePercentage = (state: TradingRootState) =>
     state.wallet.trading.settings.maxSlippagePercentage ??
