+++ suite-common/trading/src/types.ts
@@ -22,20 +22,17 @@ import type {
 import { CountryCode } from '@suite-common/geolocation';
 import { ExtendedMessageDescriptor } from '@suite-common/intl-types';
 import { AccountType, Network, NetworkSymbolExtended } from '@suite-common/wallet-config';
-import type { AccountsRootState, DeviceRootState } from '@suite-common/wallet-core';
 import {
     Account,
     BaseCurrencyOption,
     FormState,
     GeneralPrecomposedTransactionFinal,
-    SelectedAccountStatus,
 } from '@suite-common/wallet-types';
 import { PROTO, Success, Unsuccessful } from '@trezor/connect';
 import { Timer } from '@trezor/react-utils';
 import { PrimitiveType } from '@trezor/type-utils';
 
 import * as constants from './constants';
-import type { TradingState } from './reducers/tradingReducer';
 
 export type InvityServerEnvironment = 'production' | 'staging' | 'dev' | 'localhost';
 export type InvityServers = Record<InvityServerEnvironment, string>;
@@ -337,22 +334,5 @@ export type TradingVerifiedAddress =
           mac?: string;
       }
     | undefined;
-
-export type TradingRootState = {
-    wallet: {
-        trading: TradingState;
-    };
-};
-
-type SelectedAccountRootState = {
-    wallet: {
-        selectedAccount: SelectedAccountStatus;
-    };
-};
-
-export type TradingRootStateWithDeviceAndAccounts = TradingRootState &
-    DeviceRootState &
-    AccountsRootState &
-    SelectedAccountRootState;
 
 export type TradingFulfillValue = Success<{ txid: string }> | Unsuccessful | undefined;
